abstract out sequencer tx validation
